Make sure the main modifier is used at the topmost place possible
