+++ rules/common/src/main/kotlin/io/nlopez/compose/rules/ComposeModifierReused.kt
@@ -6,16 +6,12 @@ import io.nlopez.rules.core.ComposeKtVisitor
 import io.nlopez.rules.core.Emitter
 import io.nlopez.rules.core.util.emitsContent
 import io.nlopez.rules.core.util.findChildrenByClass
+import io.nlopez.rules.core.util.isUsingModifiers
 import io.nlopez.rules.core.util.modifierParameter
+import io.nlopez.rules.core.util.obtainAllModifierNames
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
-import org.jetbrains.kotlin.psi.KtBlockExpression
 import org.jetbrains.kotlin.psi.KtCallExpression
-import org.jetbrains.kotlin.psi.KtDotQualifiedExpression
-import org.jetbrains.kotlin.psi.KtExpression
 import org.jetbrains.kotlin.psi.KtFunction
-import org.jetbrains.kotlin.psi.KtProperty
-import org.jetbrains.kotlin.psi.KtReferenceExpression
-import org.jetbrains.kotlin.psi.KtValueArgumentName
 import org.jetbrains.kotlin.psi.psiUtil.siblings
 
 class ComposeModifierReused : ComposeKtVisitor {
@@ -66,66 +62,6 @@ class ComposeModifierReused : ComposeKtVisitor {
                 emitter.report(callExpression, ModifierShouldBeUsedOnceOnly, false)
             }
     }
-
-    private fun KtCallExpression.isUsingModifiers(modifierNames: List<String>): Boolean =
-        valueArguments.any { argument ->
-            when (val expression = argument.getArgumentExpression()) {
-                // if it's MyComposable(modifier) or similar
-                is KtReferenceExpression -> {
-                    modifierNames.contains(expression.text)
-                }
-                // if it's MyComposable(modifier.fillMaxWidth()) or similar
-                is KtDotQualifiedExpression -> {
-                    // On cases of multiple nested KtDotQualifiedExpressions (e.g. multiple chained methods)
-                    // we need to iterate until we find the start of the chain
-                    modifierNames.contains(expression.rootExpression.text)
-                }
-
-                else -> false
-            }
-        }
-
-    private val KtDotQualifiedExpression.rootExpression: KtExpression
-        get() {
-            var current: KtExpression = receiverExpression
-            while (current is KtDotQualifiedExpression) {
-                current = current.receiverExpression
-            }
-            return current
-        }
-
-    private fun KtBlockExpression.obtainAllModifierNames(initialName: String): List<String> {
-        var lastSize = 0
-        val tempModifierNames = mutableSetOf(initialName)
-        while (lastSize < tempModifierNames.size) {
-            lastSize = tempModifierNames.size
-            // Find usages in the current block (the original composable)
-            tempModifierNames += findModifierManipulations { tempModifierNames.contains(it) }
-            // Find usages in child composable blocks
-            tempModifierNames += findChildrenByClass<KtBlockExpression>()
-                .flatMap { block -> block.findModifierManipulations { tempModifierNames.contains(it) } }
-        }
-        return tempModifierNames.toList()
-    }
-
-    /**
-     * Find references to modifier as a property in case they try to modify or reuse the modifier that way
-     * E.g. val modifier2 = if (X) modifier.blah() else modifier.bleh()
-     */
-    private fun KtBlockExpression.findModifierManipulations(contains: (String) -> Boolean): List<String> =
-        statements
-            .filterIsInstance<KtProperty>()
-            .flatMap { property ->
-                property.findChildrenByClass<KtReferenceExpression>()
-                    .filter { referenceExpression ->
-                        val parent = referenceExpression.parent
-                        parent !is KtCallExpression &&
-                            parent !is KtValueArgumentName &&
-                            contains(referenceExpression.text)
-                    }
-                    .map { property }
-            }
-            .mapNotNull { it.nameIdentifier?.text }
 
     companion object {
         val ModifierShouldBeUsedOnceOnly = """
